tof with 5 vl6180x, moved to lib
