+++ platformio.ini
@@ -15,6 +15,6 @@ framework = arduino
 lib_deps = 
 	adafruit/Adafruit Unified Sensor@^1.1.7
 	adafruit/Adafruit BNO055@^1.6.1
-	pololu/VL53L1X@^1.3.1
+	; pololu/VL53L1X@^1.3.1
 	pololu/VL6180X@^1.3.1
 	adafruit/Adafruit NeoPixel@^1.10.7
